chore: moved Header into presentational folder
