+++ src/components/index.js
@@ -2,7 +2,7 @@
 Export components built within this directory from this file,
 so it can be picked up by this repos top level index.js
 */
-import Header from "./Header"
+import Header from "./presentational/Header"
 // import Footer from "./Footer"
 // import NoteForm from "./NoteForm"
 // import Note from "./Note"
